refactoring gait classification to the context module
