+++ src/skdh/context/meson.build
@@ -2,6 +2,7 @@ py3.install_sources(
     [
         '__init__.py',
         'core.py',
+        'gait_classification.py',
     ],
     pure: false,
     subdir: 'skdh/context',
