+++ src/skdh/context/model/lgbm_gait_classifier_no-stairs_20hz.lgbm
