+++ src/skdh/context/model/lgbm_gait_classifier_no-stairs_50hz.lgbm
