+++ src/skdh/context/model/meson.build
@@ -2,6 +2,9 @@ py3.install_sources(
     [
         '__init__.py',
         'ambulation_model.txt',
+        'gait_final_features.json',
+        'lgbm_gait_classifier_no-stairs_20hz.lgbm',
+        'lgbm_gait_classifier_no-stairs_50hz.lgbm',
     ],
     pure: false,
     subdir: 'skdh/context/model',
