+++ src/skdh/gait/meson.build
@@ -2,13 +2,11 @@ py3.install_sources(
     [
         '__init__.py',
         'core.py',
-        'gait_classification.py',
         'utility.py',
     ],
     pure: false,
     subdir: 'skdh/gait',
 )
 
-subdir('model')
 subdir('substeps')
 subdir('gait_metrics')
