--- src/skdh/gait/model/meson.build
@@ -1,10 +1,0 @@
-py3.install_sources(
-    [
-        '__init__.py',
-        'final_features.json',
-        'lgbm_gait_classifier_no-stairs_20hz.lgbm',
-        'lgbm_gait_classifier_no-stairs_50hz.lgbm',
-    ],
-    pure: false,
-    subdir: 'skdh/gait/model',
-)
